new functionality to cache Web service requests
it enables Qanary component developers to control if
equal requests to Web services should be executed or
if the result are retrieved from a cache

No implementation is required to use the cache. See
the classes RestTemplateWithCaching and CacheConfig
for the configuration.

Even if a Web service request is executed via an
instance of RestTemplateWithCaching, the cache is
only used after defining the configuration property
named `qanary.webservicecalls.cache.specs`.