qanary.commons 3.11.3: hardened Virtuoso connector
The Virtuoso Jena library shows connection problems from time to time.
Therefore, a timeout was added which is followed by re-connection approaches.
The timeout can be defined by the external property `virtuoso.query.timeout`. Default is 10 seconds.

The current implementation is catching the time out and will try to establish a new connection for 2 additional times to execute the query another time.
The whole process is done for 2 times, otherwise, an exception is raised.